Updating logic to be on the Keys instead of the registry
- This makes the API more similar to the cryptography-kotlin APIs
