--- hardware-backed/processor/src/androidMain/kotlin/co/touchlab/kjwt/hardware/AndroidKeystoreJwtProcessorRegistry.kt
@@ -1,694 +1,0 @@
-package co.touchlab.kjwt.hardware
-
-import android.os.Build
-import android.security.keystore.KeyGenParameterSpec
-import android.security.keystore.KeyProperties
-import android.security.keystore.StrongBoxUnavailableException
-import co.touchlab.kjwt.annotations.InternalKJWTApi
-import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
-import co.touchlab.kjwt.model.algorithm.EncryptionContentAlgorithm
-import co.touchlab.kjwt.model.algorithm.Jwa
-import co.touchlab.kjwt.model.algorithm.JweEncryptResult
-import co.touchlab.kjwt.model.algorithm.JwtDigest
-import co.touchlab.kjwt.model.algorithm.SigningAlgorithm
-import co.touchlab.kjwt.model.registry.JwtProcessorRegistry
-import co.touchlab.kjwt.processor.BaseJweProcessor
-import co.touchlab.kjwt.processor.BaseJwsProcessor
-import co.touchlab.kjwt.processor.JweProcessor
-import co.touchlab.kjwt.processor.JwsProcessor
-import java.security.KeyPairGenerator
-import java.security.KeyStore
-import java.security.MessageDigest
-import java.security.SecureRandom
-import java.security.Signature
-import java.security.spec.ECGenParameterSpec
-import java.security.spec.MGF1ParameterSpec
-import java.security.spec.PSSParameterSpec
-import javax.crypto.Cipher
-import javax.crypto.KeyGenerator
-import javax.crypto.Mac
-import javax.crypto.spec.GCMParameterSpec
-import javax.crypto.spec.IvParameterSpec
-import javax.crypto.spec.OAEPParameterSpec
-import javax.crypto.spec.PSource
-import javax.crypto.spec.SecretKeySpec
-
-private const val DEFAULT_JWS_KEY_ALIAS = "__kjwt_default_jws_%s_key__"
-private const val DEFAULT_JWE_KEY_ALIAS = "__kjwt_default_jwe_%s_key__"
-
-public class AndroidKeystoreJwtProcessorRegistry(
-    private val keyGenerationArguments: KeyGenerationArguments? = KeyGenerationArguments(),
-) : JwtProcessorRegistry {
-    /**
-     * Controls whether keys generated by this registry are backed by Android StrongBox.
-     *
-     * StrongBox is a dedicated secure element (separate CPU, storage, and true RNG) available on
-     * some devices running API 28+. It provides stronger security guarantees than the default
-     * Trusted Execution Environment (TEE).
-     *
-     * @see KeyGenerationArguments.strongBoxMode
-     */
-    public enum class StrongBoxMode {
-        /**
-         * Only generate keys inside StrongBox. Key generation fails if the device does not have
-         * StrongBox or runs below API 28.
-         */
-        Required,
-
-        /**
-         * Use StrongBox if available on this device (API 28+), otherwise fall back to the default
-         * TEE-backed keystore silently.
-         */
-        Preferred,
-
-        /** Do not request StrongBox-backed key storage. Default behaviour. */
-        None,
-    }
-
-    public class KeyGenerationArguments(
-        public val keySizeInBits: Int = 2048,
-        public val strongBoxMode: StrongBoxMode = StrongBoxMode.None,
-    )
-
-    @InternalKJWTApi
-    override var delegateKeyRegistry: JwtProcessorRegistry? = null
-
-    private val keystore: KeyStore
-        get() = KeyStore.getInstance("AndroidKeyStore").apply { load(null) }
-
-    override fun delegateTo(other: JwtProcessorRegistry) {
-        var cursor: JwtProcessorRegistry? = other
-        while (cursor != null) {
-            require(cursor !== this) { "Cyclic delegation detected" }
-            @OptIn(InternalKJWTApi::class)
-            cursor = cursor.delegateKeyRegistry
-        }
-        delegateKeyRegistry = other
-    }
-
-    override fun findBestJwsProcessor(
-        algorithm: SigningAlgorithm,
-        keyId: String?,
-    ): BaseJwsProcessor? {
-        if (algorithm is SigningAlgorithm.None) return null
-        val entry = getKeyOrGenerateNewIfAllowed(algorithm, keyId)
-            ?: return delegateKeyRegistry?.findBestJwsProcessor(algorithm, keyId)
-        return when (algorithm) {
-            is SigningAlgorithm.MACBased -> buildHmacProcessor(algorithm, keyId, entry)
-            is SigningAlgorithm.PKCS1Based -> buildRsaProcessor(algorithm, keyId, entry)
-            is SigningAlgorithm.PSSBased -> buildPssProcessor(algorithm, keyId, entry)
-            is SigningAlgorithm.ECDSABased -> buildEcdsaProcessor(algorithm, keyId, entry)
-            SigningAlgorithm.None -> SigningAlgorithm.None.SimpleProcessor
-        }
-    }
-
-    override fun findBestJweProcessor(
-        algorithm: EncryptionAlgorithm,
-        keyId: String?,
-    ): BaseJweProcessor? {
-        if (algorithm is EncryptionAlgorithm.Dir) {
-            return delegateKeyRegistry?.findBestJweProcessor(algorithm, keyId)
-        }
-        val entry = getEncryptionKeyOrGenerateIfAllowed(algorithm, keyId)
-            ?: return delegateKeyRegistry?.findBestJweProcessor(algorithm, keyId)
-        return buildOaepJweProcessor(algorithm as EncryptionAlgorithm.OAEPBased, keyId, entry)
-    }
-
-    // region JWS processor builders
-
-    private fun buildHmacProcessor(
-        algorithm: SigningAlgorithm.MACBased,
-        keyId: String?,
-        entry: KeyStore.Entry,
-    ): JwsProcessor {
-        val secretKey = (entry as KeyStore.SecretKeyEntry).secretKey
-        val macAlgorithm = algorithm.toJcaSignatureName()
-        return object : JwsProcessor {
-            override val algorithm: SigningAlgorithm = algorithm
-            override val keyId: String? = keyId
-
-            override suspend fun sign(data: ByteArray): ByteArray =
-                Mac.getInstance(macAlgorithm).also { it.init(secretKey) }.doFinal(data)
-
-            override suspend fun verify(data: ByteArray, signature: ByteArray): Boolean {
-                val computed = sign(data)
-                return MessageDigest.isEqual(computed, signature)
-            }
-        }
-    }
-
-    private fun buildRsaProcessor(
-        algorithm: SigningAlgorithm.PKCS1Based,
-        keyId: String?,
-        entry: KeyStore.Entry,
-    ): JwsProcessor {
-        val privateKeyEntry = entry as KeyStore.PrivateKeyEntry
-        val jcaAlgorithm = algorithm.toJcaSignatureName()
-        return object : JwsProcessor {
-            override val algorithm: SigningAlgorithm = algorithm
-            override val keyId: String? = keyId
-
-            override suspend fun sign(data: ByteArray): ByteArray =
-                Signature.getInstance(jcaAlgorithm).run {
-                    initSign(privateKeyEntry.privateKey)
-                    update(data)
-                    sign()
-                }
-
-            override suspend fun verify(data: ByteArray, signature: ByteArray): Boolean =
-                Signature.getInstance(jcaAlgorithm).run {
-                    initVerify(privateKeyEntry.certificate.publicKey)
-                    update(data)
-                    verify(signature)
-                }
-        }
-    }
-
-    private fun buildPssProcessor(
-        algorithm: SigningAlgorithm.PSSBased,
-        keyId: String?,
-        entry: KeyStore.Entry,
-    ): JwsProcessor {
-        require(Build.VERSION.SDK_INT >= 23) {
-            "PSS algorithms (PS256/PS384/PS512) require API 23+, current: ${Build.VERSION.SDK_INT}"
-        }
-        val privateKeyEntry = entry as KeyStore.PrivateKeyEntry
-        val jcaAlgorithm = algorithm.toJcaSignatureName()
-        val (digestName, saltLen) = when (algorithm) {
-            SigningAlgorithm.PS256 -> "SHA-256" to 32
-            SigningAlgorithm.PS384 -> "SHA-384" to 48
-            SigningAlgorithm.PS512 -> "SHA-512" to 64
-        }
-        val pssSpec = PSSParameterSpec(digestName, "MGF1", MGF1ParameterSpec(digestName), saltLen, 1)
-        return object : JwsProcessor {
-            override val algorithm: SigningAlgorithm = algorithm
-            override val keyId: String? = keyId
-
-            override suspend fun sign(data: ByteArray): ByteArray =
-                Signature.getInstance(jcaAlgorithm).run {
-                    setParameter(pssSpec)
-                    initSign(privateKeyEntry.privateKey)
-                    update(data)
-                    sign()
-                }
-
-            override suspend fun verify(data: ByteArray, signature: ByteArray): Boolean =
-                Signature.getInstance(jcaAlgorithm).run {
-                    setParameter(pssSpec)
-                    initVerify(privateKeyEntry.certificate.publicKey)
-                    update(data)
-                    verify(signature)
-                }
-        }
-    }
-
-    private fun buildEcdsaProcessor(
-        algorithm: SigningAlgorithm.ECDSABased,
-        keyId: String?,
-        entry: KeyStore.Entry,
-    ): JwsProcessor {
-        val privateKeyEntry = entry as KeyStore.PrivateKeyEntry
-        val jcaAlgorithm = algorithm.toJcaSignatureName()
-        val coordLen = coordinateLenForAlgorithm(algorithm)
-        return object : JwsProcessor {
-            override val algorithm: SigningAlgorithm = algorithm
-            override val keyId: String? = keyId
-
-            override suspend fun sign(data: ByteArray): ByteArray {
-                val derSignature = Signature.getInstance(jcaAlgorithm).run {
-                    initSign(privateKeyEntry.privateKey)
-                    update(data)
-                    sign()
-                }
-                return derSignature.ecdsaDerToP1363(coordLen)
-            }
-
-            override suspend fun verify(data: ByteArray, signature: ByteArray): Boolean {
-                val derSignature = signature.ecdsaP1363ToDer(coordLen)
-                return Signature.getInstance(jcaAlgorithm).run {
-                    initVerify(privateKeyEntry.certificate.publicKey)
-                    update(data)
-                    verify(derSignature)
-                }
-            }
-        }
-    }
-
-    // endregion
-
-    // region JWE processor builders
-
-    private fun buildOaepJweProcessor(
-        algorithm: EncryptionAlgorithm.OAEPBased,
-        keyId: String?,
-        entry: KeyStore.Entry,
-    ): JweProcessor {
-        val privateKeyEntry = entry as KeyStore.PrivateKeyEntry
-        val cipherName = algorithm.toOaepCipherName()
-        val oaepSpec = algorithm.toOaepParameterSpec()
-        return object : JweProcessor {
-            override val algorithm: EncryptionAlgorithm = algorithm
-            override val keyId: String? = keyId
-
-            override suspend fun encrypt(
-                data: ByteArray,
-                aad: ByteArray,
-                contentAlgorithm: EncryptionContentAlgorithm,
-            ): JweEncryptResult {
-                val cek = generateCek(contentAlgorithm)
-                val encryptedKey = Cipher.getInstance(cipherName).run {
-                    init(Cipher.ENCRYPT_MODE, privateKeyEntry.certificate.publicKey, oaepSpec)
-                    doFinal(cek)
-                }
-                val (iv, ciphertext, tag) = encryptContentJca(contentAlgorithm, cek, data, aad)
-                return JweEncryptResult(encryptedKey, iv, ciphertext, tag)
-            }
-
-            override suspend fun decrypt(
-                aad: ByteArray,
-                encryptedKey: ByteArray,
-                iv: ByteArray,
-                data: ByteArray,
-                tag: ByteArray,
-                contentAlgorithm: EncryptionContentAlgorithm,
-            ): ByteArray {
-                val cek = Cipher.getInstance(cipherName).run {
-                    init(Cipher.DECRYPT_MODE, privateKeyEntry.privateKey, oaepSpec)
-                    doFinal(encryptedKey)
-                }
-                return decryptContentJca(contentAlgorithm, cek, iv, data, tag, aad)
-            }
-        }
-    }
-
-    private fun getKeyOrGenerateNewIfAllowed(
-        algorithm: SigningAlgorithm,
-        keyId: String?,
-    ): KeyStore.Entry? {
-        val alias = keyId ?: DEFAULT_JWS_KEY_ALIAS.format(algorithm.id)
-        if (!keystore.containsAlias(alias) && keyGenerationArguments != null) {
-            createSigningKey(algorithm, keyId, keyGenerationArguments.keySizeInBits)
-        }
-        return keystore.getEntry(alias, null)
-    }
-
-    private fun getEncryptionKeyOrGenerateIfAllowed(
-        algorithm: EncryptionAlgorithm,
-        keyId: String?,
-    ): KeyStore.Entry? {
-        val alias = keyId ?: DEFAULT_JWE_KEY_ALIAS.format(algorithm.id)
-        if (!keystore.containsAlias(alias) && keyGenerationArguments != null) {
-            createEncryptionKey(algorithm, keyId, keyGenerationArguments.keySizeInBits)
-        }
-        return keystore.getEntry(alias, null)
-    }
-
-    private fun createSigningKey(
-        algorithm: SigningAlgorithm,
-        keyId: String?,
-        keySize: Int,
-    ) {
-        val alias = keyId ?: DEFAULT_JWS_KEY_ALIAS.format(algorithm.id)
-        withStrongBoxConfiguration { useStrongBox ->
-            val spec = KeyGenParameterSpec.Builder(alias, algorithm.purpose())
-                .configureFor(algorithm, keySize, useStrongBox)
-                .build()
-
-            when (algorithm) {
-                is SigningAlgorithm.MACBased -> {
-                    KeyGenerator.getInstance(algorithm.toKeyPropertiesAlgorithm(), "AndroidKeyStore")
-                        .apply { init(spec) }
-                        .generateKey()
-                }
-
-                else -> {
-                    KeyPairGenerator.getInstance(algorithm.toKeyPropertiesAlgorithm(), "AndroidKeyStore")
-                        .apply { initialize(spec) }
-                        .generateKeyPair()
-                }
-            }
-        }
-    }
-
-    private fun createEncryptionKey(
-        algorithm: EncryptionAlgorithm,
-        keyId: String?,
-        keySize: Int,
-    ) {
-        val alias = keyId ?: DEFAULT_JWE_KEY_ALIAS.format(algorithm.id)
-        withStrongBoxConfiguration { useStrongBox ->
-            val spec =
-                KeyGenParameterSpec.Builder(alias, KeyProperties.PURPOSE_ENCRYPT or KeyProperties.PURPOSE_DECRYPT)
-                    .configureFor(algorithm, keySize, useStrongBox)
-                    .build()
-
-            KeyPairGenerator.getInstance(KeyProperties.KEY_ALGORITHM_RSA, "AndroidKeyStore")
-                .apply { initialize(spec) }
-                .generateKeyPair()
-        }
-    }
-
-    private fun withStrongBoxConfiguration(
-        handler: (useStrongBox: Boolean) -> Unit,
-    ) {
-        val mode = keyGenerationArguments?.strongBoxMode ?: StrongBoxMode.None
-
-        when (mode) {
-            StrongBoxMode.None -> {
-                handler(false)
-            }
-
-            StrongBoxMode.Required -> {
-                check(Build.VERSION.SDK_INT >= 28) {
-                    "StrongBox requires API 28+, current: ${Build.VERSION.SDK_INT}"
-                }
-                handler(true)
-            }
-
-            StrongBoxMode.Preferred -> {
-                if (Build.VERSION.SDK_INT >= 28) {
-                    try {
-                        handler(true)
-                    } catch (_: StrongBoxUnavailableException) {
-                        handler(false)
-                    }
-                } else {
-                    handler(false)
-                }
-            }
-        }
-    }
-
-    // endregion
-
-    // region Algorithm mappings
-
-    private fun SigningAlgorithm.purpose(): Int = when (this) {
-        is SigningAlgorithm.MACBased -> KeyProperties.PURPOSE_SIGN
-        else -> KeyProperties.PURPOSE_SIGN or KeyProperties.PURPOSE_VERIFY
-    }
-
-    private fun SigningAlgorithm.toKeyPropertiesAlgorithm(): String = when (this) {
-        SigningAlgorithm.HS256 -> KeyProperties.KEY_ALGORITHM_HMAC_SHA256
-        SigningAlgorithm.HS384 -> KeyProperties.KEY_ALGORITHM_HMAC_SHA384
-        SigningAlgorithm.HS512 -> KeyProperties.KEY_ALGORITHM_HMAC_SHA512
-        is SigningAlgorithm.PKCS1Based -> KeyProperties.KEY_ALGORITHM_RSA
-        is SigningAlgorithm.PSSBased -> KeyProperties.KEY_ALGORITHM_RSA
-        is SigningAlgorithm.ECDSABased -> KeyProperties.KEY_ALGORITHM_EC
-        SigningAlgorithm.None -> error("None algorithm should be handled outside")
-    }
-
-    private fun SigningAlgorithm.toJcaSignatureName(): String = when (this) {
-        SigningAlgorithm.HS256 -> "HmacSHA256"
-        SigningAlgorithm.HS384 -> "HmacSHA384"
-        SigningAlgorithm.HS512 -> "HmacSHA512"
-        SigningAlgorithm.RS256 -> "SHA256withRSA"
-        SigningAlgorithm.RS384 -> "SHA384withRSA"
-        SigningAlgorithm.RS512 -> "SHA512withRSA"
-        SigningAlgorithm.PS256 -> "SHA256withRSA/PSS"
-        SigningAlgorithm.PS384 -> "SHA384withRSA/PSS"
-        SigningAlgorithm.PS512 -> "SHA512withRSA/PSS"
-        SigningAlgorithm.ES256 -> "SHA256withECDSA"
-        SigningAlgorithm.ES384 -> "SHA384withECDSA"
-        SigningAlgorithm.ES512 -> "SHA512withECDSA"
-        SigningAlgorithm.None -> error("Unsupported algorithm")
-    }
-
-    private fun EncryptionAlgorithm.toOaepCipherName(): String = when (this) {
-        EncryptionAlgorithm.RsaOaep -> "RSA/ECB/OAEPWithSHA-1AndMGF1Padding"
-        EncryptionAlgorithm.RsaOaep256 -> "RSA/ECB/OAEPWithSHA-256AndMGF1Padding"
-        EncryptionAlgorithm.Dir -> error("Dir does not use a cipher")
-    }
-
-    private fun EncryptionAlgorithm.toOaepParameterSpec(): OAEPParameterSpec = when (this) {
-        EncryptionAlgorithm.RsaOaep -> OAEPParameterSpec("SHA-1", "MGF1", MGF1ParameterSpec.SHA1, PSource.PSpecified.DEFAULT)
-        EncryptionAlgorithm.RsaOaep256 -> {
-            // setMgf1Digests (API 35+) controls the allowed MGF1 digest on the key; below API 35
-            // the keystore only permits SHA-1 for MGF1, so the spec must match accordingly.
-            val mgf1Spec = if (Build.VERSION.SDK_INT >= 35) MGF1ParameterSpec.SHA256 else MGF1ParameterSpec.SHA1
-            OAEPParameterSpec("SHA-256", "MGF1", mgf1Spec, PSource.PSpecified.DEFAULT)
-        }
-        EncryptionAlgorithm.Dir -> error("Dir does not use OAEP")
-    }
-
-    private fun KeyGenParameterSpec.Builder.configureFor(
-        algorithm: SigningAlgorithm,
-        keySize: Int,
-        useStrongBox: Boolean,
-    ) = apply {
-        if (algorithm is Jwa.UsesHashingAlgorithm) {
-            setDigests(
-                when (algorithm.digest) {
-                    JwtDigest.SHA1 -> KeyProperties.DIGEST_SHA1
-                    JwtDigest.SHA256 -> KeyProperties.DIGEST_SHA256
-                    JwtDigest.SHA384 -> KeyProperties.DIGEST_SHA384
-                    JwtDigest.SHA512 -> KeyProperties.DIGEST_SHA512
-                }
-            )
-        }
-
-        if (algorithm is SigningAlgorithm.PKCS1Based) {
-            setSignaturePaddings(KeyProperties.SIGNATURE_PADDING_RSA_PKCS1)
-            setKeySize(keySize)
-        }
-
-        if (algorithm is SigningAlgorithm.PSSBased) {
-            setSignaturePaddings(KeyProperties.SIGNATURE_PADDING_RSA_PSS)
-            setKeySize(keySize)
-        }
-
-        if (algorithm is SigningAlgorithm.ECDSABased) {
-            setAlgorithmParameterSpec(
-                ECGenParameterSpec(
-                    when (algorithm) {
-                        SigningAlgorithm.ES256 -> "secp256r1"
-                        SigningAlgorithm.ES384 -> "secp384r1"
-                        SigningAlgorithm.ES512 -> "secp521r1"
-                    }
-                )
-            )
-        }
-
-        if (useStrongBox && Build.VERSION.SDK_INT >= 28) {
-            setIsStrongBoxBacked(true)
-        }
-    }
-
-    private fun KeyGenParameterSpec.Builder.configureFor(
-        algorithm: EncryptionAlgorithm,
-        keySize: Int,
-        useStrongBox: Boolean,
-    ) = apply {
-        if (algorithm is Jwa.UsesHashingAlgorithm) {
-            setDigests(
-                when (algorithm.digest) {
-                    JwtDigest.SHA1 -> KeyProperties.DIGEST_SHA1
-                    JwtDigest.SHA256 -> KeyProperties.DIGEST_SHA256
-                    JwtDigest.SHA384 -> KeyProperties.DIGEST_SHA384
-                    JwtDigest.SHA512 -> KeyProperties.DIGEST_SHA512
-                }
-            )
-        }
-
-        if (algorithm is EncryptionAlgorithm.OAEPBased) {
-            setEncryptionPaddings(KeyProperties.ENCRYPTION_PADDING_RSA_OAEP)
-            setKeySize(keySize)
-            // setMgf1Digests was added in API 35; on older APIs SHA-1 is the only allowed MGF1 digest
-            if (algorithm == EncryptionAlgorithm.RsaOaep256 && Build.VERSION.SDK_INT >= 35) {
-                setMgf1Digests(KeyProperties.DIGEST_SHA256)
-            }
-        }
-
-        if (useStrongBox && Build.VERSION.SDK_INT >= 28) {
-            setIsStrongBoxBacked(true)
-        }
-    }
-    // endregion
-}
-
-// region ECDSA DER <-> P1363 conversion
-
-private fun coordinateLenForAlgorithm(algorithm: SigningAlgorithm.ECDSABased): Int = when (algorithm) {
-    SigningAlgorithm.ES256 -> 32
-    SigningAlgorithm.ES384 -> 48
-    SigningAlgorithm.ES512 -> 66 // P-521 is 521 bits = 66 bytes
-}
-
-/**
- * Converts a DER-encoded ECDSA signature (as returned by Android Keystore) to the
- * P1363 (IEEE 1363) raw `r || s` format required by JWT.
- */
-private fun ByteArray.ecdsaDerToP1363(coordinateLen: Int): ByteArray {
-    var pos = 0
-    // SEQUENCE tag
-    require(this[pos++] == 0x30.toByte()) { "Expected SEQUENCE tag 0x30" }
-    // Skip SEQUENCE length
-    val (_, seqLenBytes) = readDerLength(this, pos)
-    pos += seqLenBytes
-    // r INTEGER
-    require(this[pos++] == 0x02.toByte()) { "Expected INTEGER tag 0x02 for r" }
-    val (rLen, rLenBytes) = readDerLength(this, pos)
-    pos += rLenBytes
-    val rBytes = copyOfRange(pos, pos + rLen)
-    pos += rLen
-    // s INTEGER
-    require(this[pos++] == 0x02.toByte()) { "Expected INTEGER tag 0x02 for s" }
-    val (sLen, sLenBytes) = readDerLength(this, pos)
-    pos += sLenBytes
-    val sBytes = copyOfRange(pos, pos + sLen)
-
-    return stripAndPad(rBytes, coordinateLen) + stripAndPad(sBytes, coordinateLen)
-}
-
-/**
- * Converts a P1363 (raw `r || s`) ECDSA signature to DER encoding for use with JCA `Signature.verify()`.
- */
-private fun ByteArray.ecdsaP1363ToDer(coordinateLen: Int): ByteArray {
-    val r = copyOf(coordinateLen)
-    val s = copyOfRange(coordinateLen, coordinateLen * 2)
-
-    fun encodeDerInteger(raw: ByteArray): ByteArray {
-        var start = 0
-        while (start < raw.size - 1 && raw[start] == 0x00.toByte()) start++
-        val stripped = raw.copyOfRange(start, raw.size)
-        val withSign = if (stripped[0].toInt() and 0x80 != 0) byteArrayOf(0x00) + stripped else stripped
-        return byteArrayOf(0x02.toByte(), withSign.size.toByte()) + withSign
-    }
-
-    val rDer = encodeDerInteger(r)
-    val sDer = encodeDerInteger(s)
-    val body = rDer + sDer
-
-    val seqLen = if (body.size < 0x80) {
-        byteArrayOf(body.size.toByte())
-    } else {
-        byteArrayOf(0x81.toByte(), body.size.toByte())
-    }
-
-    return byteArrayOf(0x30.toByte()) + seqLen + body
-}
-
-/** Returns (length, bytesConsumed) for a DER length field. */
-private fun readDerLength(bytes: ByteArray, pos: Int): Pair<Int, Int> {
-    val first = bytes[pos].toInt() and 0xFF
-    return if (first < 0x80) {
-        Pair(first, 1)
-    } else {
-        val numOctets = first and 0x7F
-        var len = 0
-        for (i in 0 until numOctets) {
-            len = (len shl 8) or (bytes[pos + 1 + i].toInt() and 0xFF)
-        }
-        Pair(len, 1 + numOctets)
-    }
-}
-
-/** Strips the leading 0x00 DER sign byte and zero-pads to [targetLen] on the left. */
-private fun stripAndPad(bytes: ByteArray, targetLen: Int): ByteArray {
-    var start = 0
-    while (start < bytes.size - 1 && bytes[start] == 0x00.toByte()) start++
-    val stripped = bytes.copyOfRange(start, bytes.size)
-    return ByteArray(maxOf(0, targetLen - stripped.size)) + stripped
-}
-
-// endregion
-
-// region JCA content encryption helpers
-
-private val secureRandom = SecureRandom()
-
-private fun generateCek(contentAlgorithm: EncryptionContentAlgorithm): ByteArray {
-    val size = when (contentAlgorithm) {
-        EncryptionContentAlgorithm.A128GCM -> 16
-        EncryptionContentAlgorithm.A192GCM -> 24
-        EncryptionContentAlgorithm.A256GCM -> 32
-        EncryptionContentAlgorithm.A128CbcHs256 -> 32
-        EncryptionContentAlgorithm.A192CbcHs384 -> 48
-        EncryptionContentAlgorithm.A256CbcHs512 -> 64
-    }
-    return ByteArray(size).also { secureRandom.nextBytes(it) }
-}
-
-/** Returns (iv, ciphertext, tag). */
-private fun encryptContentJca(
-    contentAlgorithm: EncryptionContentAlgorithm,
-    cek: ByteArray,
-    plaintext: ByteArray,
-    aad: ByteArray,
-): Triple<ByteArray, ByteArray, ByteArray> =
-    when (contentAlgorithm) {
-        is EncryptionContentAlgorithm.AesGCMBased -> {
-            val iv = ByteArray(12).also { secureRandom.nextBytes(it) }
-            val cipher = Cipher.getInstance("AES/GCM/NoPadding")
-            cipher.init(Cipher.ENCRYPT_MODE, SecretKeySpec(cek, "AES"), GCMParameterSpec(128, iv))
-            cipher.updateAAD(aad)
-            val combined = cipher.doFinal(plaintext)
-            val ctLen = combined.size - 16
-            Triple(iv, combined.copyOf(ctLen), combined.copyOfRange(ctLen, combined.size))
-        }
-
-        is EncryptionContentAlgorithm.AesCBCBased -> {
-            val half = cek.size / 2
-            val macKey = cek.copyOf(half)
-            val encKey = cek.copyOfRange(half, cek.size)
-            val iv = ByteArray(16).also { secureRandom.nextBytes(it) }
-            val cipher = Cipher.getInstance("AES/CBC/PKCS5Padding")
-            cipher.init(Cipher.ENCRYPT_MODE, SecretKeySpec(encKey, "AES"), IvParameterSpec(iv))
-            val ciphertext = cipher.doFinal(plaintext)
-            val tag = computeCbcHmacTagJca(contentAlgorithm, macKey, aad, iv, ciphertext)
-            Triple(iv, ciphertext, tag)
-        }
-    }
-
-private fun decryptContentJca(
-    contentAlgorithm: EncryptionContentAlgorithm,
-    cek: ByteArray,
-    iv: ByteArray,
-    ciphertext: ByteArray,
-    tag: ByteArray,
-    aad: ByteArray,
-): ByteArray =
-    when (contentAlgorithm) {
-        is EncryptionContentAlgorithm.AesGCMBased -> {
-            val cipher = Cipher.getInstance("AES/GCM/NoPadding")
-            cipher.init(Cipher.DECRYPT_MODE, SecretKeySpec(cek, "AES"), GCMParameterSpec(128, iv))
-            cipher.updateAAD(aad)
-            cipher.doFinal(ciphertext + tag)
-        }
-
-        is EncryptionContentAlgorithm.AesCBCBased -> {
-            val half = cek.size / 2
-            val macKey = cek.copyOf(half)
-            val encKey = cek.copyOfRange(half, cek.size)
-            val expectedTag = computeCbcHmacTagJca(contentAlgorithm, macKey, aad, iv, ciphertext)
-            require(MessageDigest.isEqual(expectedTag, tag)) { "JWE authentication tag verification failed" }
-            val cipher = Cipher.getInstance("AES/CBC/PKCS5Padding")
-            cipher.init(Cipher.DECRYPT_MODE, SecretKeySpec(encKey, "AES"), IvParameterSpec(iv))
-            cipher.doFinal(ciphertext)
-        }
-    }
-
-/** RFC 7516 §5.2.2.1: HMAC tag for AES-CBC content encryption. */
-private fun computeCbcHmacTagJca(
-    contentAlgorithm: EncryptionContentAlgorithm.AesCBCBased,
-    macKey: ByteArray,
-    aad: ByteArray,
-    iv: ByteArray,
-    ciphertext: ByteArray,
-): ByteArray {
-    val al = aad.size.toLong() * 8
-    val alBytes = ByteArray(8) { i -> ((al shr (56 - i * 8)) and 0xFF).toByte() }
-    val macInput = aad + iv + ciphertext + alBytes
-
-    val (jcaMacAlg, tagLen) = when (contentAlgorithm) {
-        EncryptionContentAlgorithm.A128CbcHs256 -> "HmacSHA256" to 16
-        EncryptionContentAlgorithm.A192CbcHs384 -> "HmacSHA384" to 24
-        EncryptionContentAlgorithm.A256CbcHs512 -> "HmacSHA512" to 32
-    }
-
-    val mac = Mac.getInstance(jcaMacAlg)
-    mac.init(SecretKeySpec(macKey, jcaMacAlg))
-    return mac.doFinal(macInput).copyOf(tagLen)
-}
-
-// endregion
