+++ hardware-backed/processor/src/androidMain/kotlin/co/touchlab/kjwt/hardware/model/AndroidStrongBoxKeyPreference.kt
@@ -1,0 +1,66 @@
+package co.touchlab.kjwt.hardware.model
+
+import android.os.Build
+import android.security.keystore.StrongBoxUnavailableException
+
+/**
+ * Controls whether keys generated by this registry are backed by Android StrongBox.
+ *
+ * StrongBox is a dedicated secure element (separate CPU, storage, and true RNG) available on
+ * some devices running API 28+. It provides stronger security guarantees than the default
+ * Trusted Execution Environment (TEE).
+ */
+public enum class AndroidStrongBoxKeyPreference {
+    /**
+     * Only generate keys inside StrongBox. Key generation fails if the device does not have
+     * StrongBox or runs below API 28.
+     */
+    Required,
+
+    /**
+     * Use StrongBox if available on this device (API 28+), otherwise fall back to the default
+     * TEE-backed keystore silently.
+     */
+    Preferred,
+
+    /** Do not request StrongBox-backed key storage. Default behaviour. */
+    None;
+
+    /**
+     * Invokes [handler] with a boolean indicating whether the Android Keystore should be asked
+     * to back the key with StrongBox.
+     *
+     * - [None]: always calls `handler(false)`.
+     * - [Required]: asserts API 28+ is available, then calls `handler(true)`. Throws
+     *   [IllegalStateException] if the device is below API 28.
+     * - [Preferred]: calls `handler(true)` on API 28+, catching [android.security.keystore.StrongBoxUnavailableException]
+     *   and retrying with `handler(false)` if the device has no StrongBox; calls `handler(false)`
+     *   directly on older API levels.
+     */
+    public fun runWithFlag(handler: (useStrongBox: Boolean) -> Unit) {
+        when (this) {
+            None -> {
+                handler(false)
+            }
+
+            Required -> {
+                check(Build.VERSION.SDK_INT >= 28) {
+                    "StrongBox requires API 28+, current: ${Build.VERSION.SDK_INT}"
+                }
+                handler(true)
+            }
+
+            Preferred -> {
+                if (Build.VERSION.SDK_INT >= 28) {
+                    try {
+                        handler(true)
+                    } catch (_: StrongBoxUnavailableException) {
+                        handler(false)
+                    }
+                } else {
+                    handler(false)
+                }
+            }
+        }
+    }
+}
